+++ lua/CompMod/Marine/ARC/Health/Post/BalanceHealth.lua
@@ -1,0 +1,3 @@
+kARCHealth = 2200
+kARCArmor = 400
+kARCPointValue = 5